refactor: move transform streams to a separate folder
